+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added support for shell script highlighting
 - Added support for contains, minContains and maxContains constrains
 - Added support for if-then-else constrains
+- Added support for dependentSchemas constrains
 
 ### 2.9.1